Set compilation.  Refs #170.
* Reorganize set and dict internal functions used for compilation,
  for codecov coverage.
